+++ app/main/cennznet/cennzNetConfig.js
@@ -1,6 +1,7 @@
 // @flow
 import os from 'os';
 
+import appConfig from 'app/config';
 import { NetworkNameOptions } from 'common/types/cennznet-node.types';
 import type { LauncherConfig } from '../launcherConfig';
 
@@ -17,7 +18,16 @@ export const prepareArgs = (config: LauncherConfig) => {
   // default node name
   const nameArgIndex = args.findIndex(item => item === '--name');
   if (nameArgIndex < 0) {
-    args.push('--name', '❤️Odin-' + os.userInfo().username);
+    const { username } = os.userInfo();
+    const first = os.userInfo().username.substring(0, 1);
+    const last = os.userInfo().username.substring(username.length - 1);
+    let maskValue = username.substring(1, username.length - 1);
+    if(maskValue.length > 10) {
+      maskValue = '**********';
+    }
+    const mask = maskValue.replace(/\w/g,"*");
+    const maskedName = first + mask + last;
+    args.push('--name', '❤️' + appConfig.app.name + '-' + maskedName);
   }
 
   // default chain